fix: majority trivial fixed for non integers

+++ examples/trivial_autocomplete.py
@@ -54,8 +54,9 @@ def predict(train_table: Table, pred_table: Table, name: str) -> np.ndarray:
     elif name == "random":
         pred = np.random.rand(len(pred_table))
     elif name == "majority":
-        past_target = train_table.df[task.target_col].astype(int)
-        majority_label = int(past_target.mode().iloc[0])
+        # fix: not only 0/1 labels to handle but also e.g. 't', 'f'
+        past_target = train_table.df[task.target_col]
+        majority_label = past_target.mode().iloc[0]
         pred = np.full(len(pred_table), majority_label)
     else:
         raise ValueError(f"Unknown eval name called {name}.")
